New classes, lots of refactoring
Split the Keybase class in to the base Keybase class which represents read-only type actions with publicly available information about a user in the keybase.io data store.

And created a new KeybaseAdmin class that will eventually extend Keybase with administrative commands so you can add and revoke keys, logout sessions, etc. for a user. For not it's basically doing nothing as I work out how to log in with their 2-phase login scheme.

I also added a KeybasePublicKey class that represents a single bit of public key information in a public_key record for a user in the data store. It's a bit of a glorified dictionary right now, but it will get some more heavy-lifting methods in the near future as I implement encrypt and verify signature methods.