Integrated GnuPG via the gnugp-python module
The KeybasePublicKey class now creates a gnupg.GPG class instance within itself. This is the tool that will do the real, heavy lifting for us when it comes to encrypting and verifying signatures.

The class setups up a completely private keyring for the GPG class that contains one and only one key: the Keybase key associated with the KeybasePublicKey class instance. When the class instance is destroyed so is the temporary keyring. In this way the user's default keyring is never polluted.

TODO:

* Still have to write the KeybasePublicKey.encrypt() and .sign() methods