We have verification!
Signature verification that is!

You can now verify the signed data using a myriad of different methods on both the Keybase and KeybasePublicKey objects.

The documentation has been updated to provide <gasp!> actual examples that work. There was a bit of refactoring work as well and there's likely a bunch more of that to come.