A rather significant simplification to file-based signature verification
Wherein the cold, harsh light of the following evening Ian realizes that having a verify_file_detached() and verifiy_file_embedded() is kind of dumb and there should just be a verify() and verify_file() and that should  be the end.

So there's only a verify_file() now and it supports both embedded and detached signature verification with the calling semantics being the name of a file, not a file object, in all cases.