Fixes for latest gnupg and some refactoring to the library
Fix ups so everything works with the latest gnupg release from
@isislovecruft. I also refactored the Keybase main class a bit so you
can't create an instance without giving it a username (that was silly to
being with). This is prep work for the new `discover` class method
that's coming that'll let you look up users by their twitter handle or
github account name.