Implement keybase.discover, bit of refactor as well
Implement the module-level `discover()` function that looks up Keybase
users by their metadata like Twitter handle or Github user name and
returns a tuple of Keybase objects for the found users.

Refactored a bunch of stuff so it's easier to use. The
`test/test_regressions.py` is a good example file if you want to see how
it's changed. The `README.md` also contains better instructions for
using the module now.